# Compact table "neutral".