--- a/runewidth_test.go
+++ b/runewidth_test.go
@@ -72,7 +72,7 @@ func isCompact(t *testing.T, tbl table) bool {
 			t.Errorf("table invalid: %v", tbl[i])
 			return false
 		}
-		if i > 0 && tbl[i-1].last+1 >= tbl[i].first { // can be combined into one entry
+		if i+1 < len(tbl) && tbl[i].last+1 >= tbl[i+1].first { // can be combined into one entry
 			t.Errorf("table not compact: %v %v", tbl[i-1], tbl[i])
 			return false
 		}
@@ -80,16 +80,45 @@ func isCompact(t *testing.T, tbl table) bool {
 	return true
 }
 
+// This is a utility function in case that a table has changed.
+func printCompactTable(tbl table) {
+	counter := 0
+	printEntry := func(first, last rune) {
+		if counter%3 == 0 {
+			fmt.Printf("\t")
+		}
+		fmt.Printf("{0x%04X, 0x%04X},", first, last)
+		if (counter+1)%3 == 0 {
+			fmt.Printf("\n")
+		} else {
+			fmt.Printf(" ")
+		}
+		counter++
+	}
+
+	sort.Sort(&tbl) // just in case
+	first := rune(-1)
+	for i := range tbl {
+		if first < 0 {
+			first = tbl[i].first
+		}
+		if i+1 < len(tbl) && tbl[i].last+1 >= tbl[i+1].first { // can be combined into one entry
+			continue
+		}
+		printEntry(first, tbl[i].last)
+		first = -1
+	}
+	fmt.Printf("\n\n")
+}
+
 func TestSorted(t *testing.T) {
 	for _, tbl := range tables {
 		if !sort.IsSorted(&tbl) {
 			t.Errorf("table not sorted")
 		}
-		if &tbl[0] == &neutral[0] {
-			continue // TODO: compact the table "neutral"
-		}
 		if !isCompact(t, tbl) {
 			t.Errorf("table not compact")
+			// printCompactTable(tbl)
 		}
 	}
 }