Saves some space and also a small speedup.

name                old time/op  new time/op  delta
TablePrivate-4      5.82ms ± 0%  5.83ms ± 0%    ~     (p=0.690 n=5+5)
TableNonprint-4     7.08ms ± 0%  7.09ms ± 0%    ~     (p=0.151 n=5+5)
TableCombining-4    11.1ms ± 0%  11.1ms ± 0%    ~     (p=0.310 n=5+5)
TableDoublewidth-4  10.1ms ± 1%  10.0ms ± 1%    ~     (p=0.690 n=5+5)
TableAmbiguous-4    10.8ms ± 0%  10.8ms ± 0%    ~     (p=0.310 n=5+5)
TableEmoji-4        8.08ms ± 0%  8.09ms ± 0%    ~     (p=0.222 n=5+5)
TableNotassigned-4  11.0ms ± 0%  11.0ms ± 0%    ~     (p=0.310 n=5+5)
TableNeutral-4      14.4ms ± 0%  13.3ms ± 0%  -7.34%  (p=0.008 n=5+5)